ref: bearer token moved folder middlewares
