+++ src/routes/features/car.routes.js
@@ -2,7 +2,7 @@ const carController = require('../../app/controller/carController');
 const validateBody = require('../../app/validation/car/body.js');
 const validateId = require('../../app/validation/id/id.js');
 const validateFind = require('../../app/validation/car/findCar.js');
-const createToken = require('../../app/utils/Auth/bearerToken.js');
+const createToken = require('../../app/middlewares/bearerToken.js');
 
 module.exports = (server, routes, prefix = '/api/v1/car') => {
   routes.post('/', createToken, validateBody, carController.create);
